moves helper functions into own file
