+++ lambdas/serve/signUp.js
@@ -1,1 +1,1 @@
-!function(t,r){for(var n in r)t[n]=r[n]}(exports,function(t){var r={};function n(e){if(r[e])return r[e].exports;var o=r[e]={i:e,l:!1,exports:{}};return t[e].call(o.exports,o,o.exports,n),o.l=!0,o.exports}return n.m=t,n.c=r,n.d=function(t,r,e){n.o(t,r)||Object.defineProperty(t,r,{enumerable:!0,get:e})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,r){if(1&r&&(t=n(t)),8&r)return t;if(4&r&&"object"==typeof t&&t&&t.__esModule)return t;var e=Object.create(null);if(n.r(e),Object.defineProperty(e,"default",{enumerable:!0,value:t}),2&r&&"string"!=typeof t)for(var o in t)n.d(e,o,function(r){return t[r]}.bind(null,o));return e},n.n=function(t){var r=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(r,"a",r),r},n.o=function(t,r){return Object.prototype.hasOwnProperty.call(t,r)},n.p="",n(n.s=86)}([function(t,r){var n=t.exports="undefined"!=typeof window&&window.Math==Math?window:"undefined"!=typeof self&&self.Math==Math?self:Function("return this")();"number"==typeof __g&&(__g=n)},function(t,r,n){var e=n(31)("wks"),o=n(32),i=n(0).Symbol,u="function"==typeof i;(t.exports=function(t){return e[t]||(e[t]=u&&i[t]||(u?i:o)("Symbol."+t))}).store=e},function(t,r){var n=t.exports={version:"2.6.5"};"number"==typeof __e&&(__e=n)},function(t,r,n){var e=n(7);t.exports=function(t){if(!e(t))throw TypeError(t+" is not an object!");return t}},function(t,r,n){var e=n(14),o=n(28);t.exports=n(8)?function(t,r,n){return e.f(t,r,o(1,n))}:function(t,r,n){return t[r]=n,t}},function(t,r,n){t.exports=n(60)},function(t,r,n){t.exports=n(75)},function(t,r){t.exports=function(t){return"object"==typeof t?null!==t:"function"==typeof t}},function(t,r,n){t.exports=!n(27)(function(){return 7!=Object.defineProperty({},"a",{get:function(){return 7}}).a})},function(t,r){t.exports={}},function(t,r,n){var e=n(0),o=n(2),i=n(12),u=n(4),c=n(15),a=function(t,r,n){var s,f,l,p=t&a.F,h=t&a.G,v=t&a.S,y=t&a.P,d=t&a.B,m=t&a.W,g=h?o:o[r]||(o[r]={}),x=g.prototype,w=h?e:v?e[r]:(e[r]||{}).prototype;for(s in h&&(n=r),n)(f=!p&&w&&void 0!==w[s])&&c(g,s)||(l=f?w[s]:n[s],g[s]=h&&"function"!=typeof w[s]?n[s]:d&&f?i(l,e):m&&w[s]==l?function(t){var r=function(r,n,e){if(this instanceof t){switch(arguments.length){case 0:return new t;case 1:return new t(r);case 2:return new t(r,n)}return new t(r,n,e)}return t.apply(this,arguments)};return r.prototype=t.prototype,r}(l):y&&"function"==typeof l?i(Function.call,l):l,y&&((g.virtual||(g.virtual={}))[s]=l,t&a.R&&x&&!x[s]&&u(x,s,l)))};a.F=1,a.G=2,a.S=4,a.P=8,a.B=16,a.W=32,a.U=64,a.R=128,t.exports=a},function(t,r){var n={}.toString;t.exports=function(t){return n.call(t).slice(8,-1)}},function(t,r,n){var e=n(13);t.exports=function(t,r,n){if(e(t),void 0===r)return t;switch(n){case 1:return function(n){return t.call(r,n)};case 2:return function(n,e){return t.call(r,n,e)};case 3:return function(n,e,o){return t.call(r,n,e,o)}}return function(){return t.apply(r,arguments)}}},function(t,r){t.exports=function(t){if("function"!=typeof t)throw TypeError(t+" is not a function!");return t}},function(t,r,n){var e=n(3),o=n(43),i=n(44),u=Object.defineProperty;r.f=n(8)?Object.defineProperty:function(t,r,n){if(e(t),r=i(r,!0),e(n),o)try{return u(t,r,n)}catch(t){}if("get"in n||"set"in n)throw TypeError("Accessors not supported!");return"value"in n&&(t[r]=n.value),t}},function(t,r){var n={}.hasOwnProperty;t.exports=function(t,r){return n.call(t,r)}},function(t,r,n){"use strict";n.d(r,"a",function(){return u});var e=n(5),o=n.n(e);function i(t,r,n,e,i,u,c){try{var a=t[u](c),s=a.value}catch(t){return void n(t)}a.done?r(s):o.a.resolve(s).then(e,i)}function u(t){return function(){var r=this,n=arguments;return new o.a(function(e,o){var u=t.apply(r,n);function c(t){i(u,e,o,c,a,"next",t)}function a(t){i(u,e,o,c,a,"throw",t)}c(void 0)})}}},function(t,r,n){var e=n(7),o=n(0).document,i=e(o)&&e(o.createElement);t.exports=function(t){return i?o.createElement(t):{}}},function(t,r,n){var e=n(48),o=n(19);t.exports=function(t){return e(o(t))}},function(t,r){t.exports=function(t){if(null==t)throw TypeError("Can't call method on  "+t);return t}},function(t,r){t.exports=!0},function(t,r){var n=Math.ceil,e=Math.floor;t.exports=function(t){return isNaN(t=+t)?0:(t>0?e:n)(t)}},function(t,r,n){var e=n(31)("keys"),o=n(32);t.exports=function(t){return e[t]||(e[t]=o(t))}},function(t,r,n){var e=n(14).f,o=n(15),i=n(1)("toStringTag");t.exports=function(t,r,n){t&&!o(t=n?t:t.prototype,i)&&e(t,i,{configurable:!0,value:r})}},function(t,r,n){"use strict";var e=n(13);function o(t){var r,n;this.promise=new t(function(t,e){if(void 0!==r||void 0!==n)throw TypeError("Bad Promise constructor");r=t,n=e}),this.resolve=e(r),this.reject=e(n)}t.exports.f=function(t){return new o(t)}},function(t,r){t.exports=require("crypto")},function(t,r){t.exports=require("util")},function(t,r){t.exports=function(t){try{return!!t()}catch(t){return!0}}},function(t,r){t.exports=function(t,r){return{enumerable:!(1&t),configurable:!(2&t),writable:!(4&t),value:r}}},function(t,r,n){"use strict";var e=n(20),o=n(10),i=n(49),u=n(4),c=n(9),a=n(50),s=n(23),f=n(57),l=n(1)("iterator"),p=!([].keys&&"next"in[].keys()),h=function(){return this};t.exports=function(t,r,n,v,y,d,m){a(n,r,v);var g,x,w,_=function(t){if(!p&&t in S)return S[t];switch(t){case"keys":case"values":return function(){return new n(this,t)}}return function(){return new n(this,t)}},b=r+" Iterator",O="values"==y,j=!1,S=t.prototype,L=S[l]||S["@@iterator"]||y&&S[y],P=L||_(y),E=y?O?_("entries"):P:void 0,T="Array"==r&&S.entries||L;if(T&&(w=f(T.call(new t)))!==Object.prototype&&w.next&&(s(w,b,!0),e||"function"==typeof w[l]||u(w,l,h)),O&&L&&"values"!==L.name&&(j=!0,P=function(){return L.call(this)}),e&&!m||!p&&!j&&S[l]||u(S,l,P),c[r]=P,c[b]=h,y)if(g={values:O?P:_("values"),keys:d?P:_("keys"),entries:E},m)for(x in g)x in S||i(S,x,g[x]);else o(o.P+o.F*(p||j),r,g);return g}},function(t,r,n){var e=n(21),o=Math.min;t.exports=function(t){return t>0?o(e(t),9007199254740991):0}},function(t,r,n){var e=n(2),o=n(0),i=o["__core-js_shared__"]||(o["__core-js_shared__"]={});(t.exports=function(t,r){return i[t]||(i[t]=void 0!==r?r:{})})("versions",[]).push({version:e.version,mode:n(20)?"pure":"global",copyright:"© 2019 Denis Pushkarev (zloirock.ru)"})},function(t,r){var n=0,e=Math.random();t.exports=function(t){return"Symbol(".concat(void 0===t?"":t,")_",(++n+e).toString(36))}},function(t,r){t.exports="constructor,hasOwnProperty,isPrototypeOf,propertyIsEnumerable,toLocaleString,toString,valueOf".split(",")},function(t,r,n){var e=n(0).document;t.exports=e&&e.documentElement},function(t,r,n){var e=n(11),o=n(1)("toStringTag"),i="Arguments"==e(function(){return arguments}());t.exports=function(t){var r,n,u;return void 0===t?"Undefined":null===t?"Null":"string"==typeof(n=function(t,r){try{return t[r]}catch(t){}}(r=Object(t),o))?n:i?e(r):"Object"==(u=e(r))&&"function"==typeof r.callee?"Arguments":u}},function(t,r,n){var e=n(3),o=n(13),i=n(1)("species");t.exports=function(t,r){var n,u=e(t).constructor;return void 0===u||null==(n=e(u)[i])?r:o(n)}},function(t,r,n){var e,o,i,u=n(12),c=n(67),a=n(34),s=n(17),f=n(0),l=f.process,p=f.setImmediate,h=f.clearImmediate,v=f.MessageChannel,y=f.Dispatch,d=0,m={},g=function(){var t=+this;if(m.hasOwnProperty(t)){var r=m[t];delete m[t],r()}},x=function(t){g.call(t.data)};p&&h||(p=function(t){for(var r=[],n=1;arguments.length>n;)r.push(arguments[n++]);return m[++d]=function(){c("function"==typeof t?t:Function(t),r)},e(d),d},h=function(t){delete m[t]},"process"==n(11)(l)?e=function(t){l.nextTick(u(g,t,1))}:y&&y.now?e=function(t){y.now(u(g,t,1))}:v?(i=(o=new v).port2,o.port1.onmessage=x,e=u(i.postMessage,i,1)):f.addEventListener&&"function"==typeof postMessage&&!f.importScripts?(e=function(t){f.postMessage(t+"","*")},f.addEventListener("message",x,!1)):e="onreadystatechange"in s("script")?function(t){a.appendChild(s("script")).onreadystatechange=function(){a.removeChild(this),g.call(t)}}:function(t){setTimeout(u(g,t,1),0)}),t.exports={set:p,clear:h}},function(t,r){t.exports=function(t){try{return{e:!1,v:t()}}catch(t){return{e:!0,v:t}}}},function(t,r,n){var e=n(3),o=n(7),i=n(24);t.exports=function(t,r){if(e(t),o(r)&&r.constructor===t)return r;var n=i.f(t);return(0,n.resolve)(r),n.promise}},function(t,r,n){n(45);for(var e=n(0),o=n(4),i=n(9),u=n(1)("toStringTag"),c="CSSRuleList,CSSStyleDeclaration,CSSValueList,ClientRectList,DOMRectList,DOMStringList,DOMTokenList,DataTransferItemList,FileList,HTMLAllCollection,HTMLCollection,HTMLFormElement,HTMLSelectElement,MediaList,MimeTypeArray,NamedNodeMap,NodeList,PaintRequestList,Plugin,PluginArray,SVGLengthList,SVGNumberList,SVGPathSegList,SVGPointList,SVGStringList,SVGTransformList,SourceBufferList,StyleSheetList,TextTrackCueList,TextTrackList,TouchList".split(","),a=0;a<c.length;a++){var s=c[a],f=e[s],l=f&&f.prototype;l&&!l[u]&&o(l,u,s),i[s]=i.Array}},function(t,r,n){"use strict";var e=n(59)(!0);n(29)(String,"String",function(t){this._t=String(t),this._i=0},function(){var t,r=this._t,n=this._i;return n>=r.length?{value:void 0,done:!0}:(t=e(r,n),this._i+=t.length,{value:t,done:!1})})},function(t,r,n){var e=n(35),o=n(1)("iterator"),i=n(9);t.exports=n(2).getIteratorMethod=function(t){if(null!=t)return t[o]||t["@@iterator"]||i[e(t)]}},function(t,r,n){t.exports=!n(8)&&!n(27)(function(){return 7!=Object.defineProperty(n(17)("div"),"a",{get:function(){return 7}}).a})},function(t,r,n){var e=n(7);t.exports=function(t,r){if(!e(t))return t;var n,o;if(r&&"function"==typeof(n=t.toString)&&!e(o=n.call(t)))return o;if("function"==typeof(n=t.valueOf)&&!e(o=n.call(t)))return o;if(!r&&"function"==typeof(n=t.toString)&&!e(o=n.call(t)))return o;throw TypeError("Can't convert object to primitive value")}},function(t,r,n){"use strict";var e=n(46),o=n(47),i=n(9),u=n(18);t.exports=n(29)(Array,"Array",function(t,r){this._t=u(t),this._i=0,this._k=r},function(){var t=this._t,r=this._k,n=this._i++;return!t||n>=t.length?(this._t=void 0,o(1)):o(0,"keys"==r?n:"values"==r?t[n]:[n,t[n]])},"values"),i.Arguments=i.Array,e("keys"),e("values"),e("entries")},function(t,r){t.exports=function(){}},function(t,r){t.exports=function(t,r){return{value:r,done:!!t}}},function(t,r,n){var e=n(11);t.exports=Object("z").propertyIsEnumerable(0)?Object:function(t){return"String"==e(t)?t.split(""):Object(t)}},function(t,r,n){t.exports=n(4)},function(t,r,n){"use strict";var e=n(51),o=n(28),i=n(23),u={};n(4)(u,n(1)("iterator"),function(){return this}),t.exports=function(t,r,n){t.prototype=e(u,{next:o(1,n)}),i(t,r+" Iterator")}},function(t,r,n){var e=n(3),o=n(52),i=n(33),u=n(22)("IE_PROTO"),c=function(){},a=function(){var t,r=n(17)("iframe"),e=i.length;for(r.style.display="none",n(34).appendChild(r),r.src="javascript:",(t=r.contentWindow.document).open(),t.write("<script>document.F=Object<\/script>"),t.close(),a=t.F;e--;)delete a.prototype[i[e]];return a()};t.exports=Object.create||function(t,r){var n;return null!==t?(c.prototype=e(t),n=new c,c.prototype=null,n[u]=t):n=a(),void 0===r?n:o(n,r)}},function(t,r,n){var e=n(14),o=n(3),i=n(53);t.exports=n(8)?Object.defineProperties:function(t,r){o(t);for(var n,u=i(r),c=u.length,a=0;c>a;)e.f(t,n=u[a++],r[n]);return t}},function(t,r,n){var e=n(54),o=n(33);t.exports=Object.keys||function(t){return e(t,o)}},function(t,r,n){var e=n(15),o=n(18),i=n(55)(!1),u=n(22)("IE_PROTO");t.exports=function(t,r){var n,c=o(t),a=0,s=[];for(n in c)n!=u&&e(c,n)&&s.push(n);for(;r.length>a;)e(c,n=r[a++])&&(~i(s,n)||s.push(n));return s}},function(t,r,n){var e=n(18),o=n(30),i=n(56);t.exports=function(t){return function(r,n,u){var c,a=e(r),s=o(a.length),f=i(u,s);if(t&&n!=n){for(;s>f;)if((c=a[f++])!=c)return!0}else for(;s>f;f++)if((t||f in a)&&a[f]===n)return t||f||0;return!t&&-1}}},function(t,r,n){var e=n(21),o=Math.max,i=Math.min;t.exports=function(t,r){return(t=e(t))<0?o(t+r,0):i(t,r)}},function(t,r,n){var e=n(15),o=n(58),i=n(22)("IE_PROTO"),u=Object.prototype;t.exports=Object.getPrototypeOf||function(t){return t=o(t),e(t,i)?t[i]:"function"==typeof t.constructor&&t instanceof t.constructor?t.constructor.prototype:t instanceof Object?u:null}},function(t,r,n){var e=n(19);t.exports=function(t){return Object(e(t))}},function(t,r,n){var e=n(21),o=n(19);t.exports=function(t){return function(r,n){var i,u,c=String(o(r)),a=e(n),s=c.length;return a<0||a>=s?t?"":void 0:(i=c.charCodeAt(a))<55296||i>56319||a+1===s||(u=c.charCodeAt(a+1))<56320||u>57343?t?c.charAt(a):i:t?c.slice(a,a+2):u-56320+(i-55296<<10)+65536}}},function(t,r,n){n(61),n(41),n(40),n(62),n(73),n(74),t.exports=n(2).Promise},function(t,r){},function(t,r,n){"use strict";var e,o,i,u,c=n(20),a=n(0),s=n(12),f=n(35),l=n(10),p=n(7),h=n(13),v=n(63),y=n(64),d=n(36),m=n(37).set,g=n(68)(),x=n(24),w=n(38),_=n(69),b=n(39),O=a.TypeError,j=a.process,S=j&&j.versions,L=S&&S.v8||"",P=a.Promise,E="process"==f(j),T=function(){},k=o=x.f,M=!!function(){try{var t=P.resolve(1),r=(t.constructor={})[n(1)("species")]=function(t){t(T,T)};return(E||"function"==typeof PromiseRejectionEvent)&&t.then(T)instanceof r&&0!==L.indexOf("6.6")&&-1===_.indexOf("Chrome/66")}catch(t){}}(),N=function(t){var r;return!(!p(t)||"function"!=typeof(r=t.then))&&r},R=function(t,r){if(!t._n){t._n=!0;var n=t._c;g(function(){for(var e=t._v,o=1==t._s,i=0,u=function(r){var n,i,u,c=o?r.ok:r.fail,a=r.resolve,s=r.reject,f=r.domain;try{c?(o||(2==t._h&&C(t),t._h=1),!0===c?n=e:(f&&f.enter(),n=c(e),f&&(f.exit(),u=!0)),n===r.promise?s(O("Promise-chain cycle")):(i=N(n))?i.call(n,a,s):a(n)):s(e)}catch(t){f&&!u&&f.exit(),s(t)}};n.length>i;)u(n[i++]);t._c=[],t._n=!1,r&&!t._h&&A(t)})}},A=function(t){m.call(a,function(){var r,n,e,o=t._v,i=F(t);if(i&&(r=w(function(){E?j.emit("unhandledRejection",o,t):(n=a.onunhandledrejection)?n({promise:t,reason:o}):(e=a.console)&&e.error&&e.error("Unhandled promise rejection",o)}),t._h=E||F(t)?2:1),t._a=void 0,i&&r.e)throw r.v})},F=function(t){return 1!==t._h&&0===(t._a||t._c).length},C=function(t){m.call(a,function(){var r;E?j.emit("rejectionHandled",t):(r=a.onrejectionhandled)&&r({promise:t,reason:t._v})})},G=function(t){var r=this;r._d||(r._d=!0,(r=r._w||r)._v=t,r._s=2,r._a||(r._a=r._c.slice()),R(r,!0))},I=function(t){var r,n=this;if(!n._d){n._d=!0,n=n._w||n;try{if(n===t)throw O("Promise can't be resolved itself");(r=N(t))?g(function(){var e={_w:n,_d:!1};try{r.call(t,s(I,e,1),s(G,e,1))}catch(t){G.call(e,t)}}):(n._v=t,n._s=1,R(n,!1))}catch(t){G.call({_w:n,_d:!1},t)}}};M||(P=function(t){v(this,P,"Promise","_h"),h(t),e.call(this);try{t(s(I,this,1),s(G,this,1))}catch(t){G.call(this,t)}},(e=function(t){this._c=[],this._a=void 0,this._s=0,this._d=!1,this._v=void 0,this._h=0,this._n=!1}).prototype=n(70)(P.prototype,{then:function(t,r){var n=k(d(this,P));return n.ok="function"!=typeof t||t,n.fail="function"==typeof r&&r,n.domain=E?j.domain:void 0,this._c.push(n),this._a&&this._a.push(n),this._s&&R(this,!1),n.promise},catch:function(t){return this.then(void 0,t)}}),i=function(){var t=new e;this.promise=t,this.resolve=s(I,t,1),this.reject=s(G,t,1)},x.f=k=function(t){return t===P||t===u?new i(t):o(t)}),l(l.G+l.W+l.F*!M,{Promise:P}),n(23)(P,"Promise"),n(71)("Promise"),u=n(2).Promise,l(l.S+l.F*!M,"Promise",{reject:function(t){var r=k(this);return(0,r.reject)(t),r.promise}}),l(l.S+l.F*(c||!M),"Promise",{resolve:function(t){return b(c&&this===u?P:this,t)}}),l(l.S+l.F*!(M&&n(72)(function(t){P.all(t).catch(T)})),"Promise",{all:function(t){var r=this,n=k(r),e=n.resolve,o=n.reject,i=w(function(){var n=[],i=0,u=1;y(t,!1,function(t){var c=i++,a=!1;n.push(void 0),u++,r.resolve(t).then(function(t){a||(a=!0,n[c]=t,--u||e(n))},o)}),--u||e(n)});return i.e&&o(i.v),n.promise},race:function(t){var r=this,n=k(r),e=n.reject,o=w(function(){y(t,!1,function(t){r.resolve(t).then(n.resolve,e)})});return o.e&&e(o.v),n.promise}})},function(t,r){t.exports=function(t,r,n,e){if(!(t instanceof r)||void 0!==e&&e in t)throw TypeError(n+": incorrect invocation!");return t}},function(t,r,n){var e=n(12),o=n(65),i=n(66),u=n(3),c=n(30),a=n(42),s={},f={};(r=t.exports=function(t,r,n,l,p){var h,v,y,d,m=p?function(){return t}:a(t),g=e(n,l,r?2:1),x=0;if("function"!=typeof m)throw TypeError(t+" is not iterable!");if(i(m)){for(h=c(t.length);h>x;x++)if((d=r?g(u(v=t[x])[0],v[1]):g(t[x]))===s||d===f)return d}else for(y=m.call(t);!(v=y.next()).done;)if((d=o(y,g,v.value,r))===s||d===f)return d}).BREAK=s,r.RETURN=f},function(t,r,n){var e=n(3);t.exports=function(t,r,n,o){try{return o?r(e(n)[0],n[1]):r(n)}catch(r){var i=t.return;throw void 0!==i&&e(i.call(t)),r}}},function(t,r,n){var e=n(9),o=n(1)("iterator"),i=Array.prototype;t.exports=function(t){return void 0!==t&&(e.Array===t||i[o]===t)}},function(t,r){t.exports=function(t,r,n){var e=void 0===n;switch(r.length){case 0:return e?t():t.call(n);case 1:return e?t(r[0]):t.call(n,r[0]);case 2:return e?t(r[0],r[1]):t.call(n,r[0],r[1]);case 3:return e?t(r[0],r[1],r[2]):t.call(n,r[0],r[1],r[2]);case 4:return e?t(r[0],r[1],r[2],r[3]):t.call(n,r[0],r[1],r[2],r[3])}return t.apply(n,r)}},function(t,r,n){var e=n(0),o=n(37).set,i=e.MutationObserver||e.WebKitMutationObserver,u=e.process,c=e.Promise,a="process"==n(11)(u);t.exports=function(){var t,r,n,s=function(){var e,o;for(a&&(e=u.domain)&&e.exit();t;){o=t.fn,t=t.next;try{o()}catch(e){throw t?n():r=void 0,e}}r=void 0,e&&e.enter()};if(a)n=function(){u.nextTick(s)};else if(!i||e.navigator&&e.navigator.standalone)if(c&&c.resolve){var f=c.resolve(void 0);n=function(){f.then(s)}}else n=function(){o.call(e,s)};else{var l=!0,p=document.createTextNode("");new i(s).observe(p,{characterData:!0}),n=function(){p.data=l=!l}}return function(e){var o={fn:e,next:void 0};r&&(r.next=o),t||(t=o,n()),r=o}}},function(t,r,n){var e=n(0).navigator;t.exports=e&&e.userAgent||""},function(t,r,n){var e=n(4);t.exports=function(t,r,n){for(var o in r)n&&t[o]?t[o]=r[o]:e(t,o,r[o]);return t}},function(t,r,n){"use strict";var e=n(0),o=n(2),i=n(14),u=n(8),c=n(1)("species");t.exports=function(t){var r="function"==typeof o[t]?o[t]:e[t];u&&r&&!r[c]&&i.f(r,c,{configurable:!0,get:function(){return this}})}},function(t,r,n){var e=n(1)("iterator"),o=!1;try{var i=[7][e]();i.return=function(){o=!0},Array.from(i,function(){throw 2})}catch(t){}t.exports=function(t,r){if(!r&&!o)return!1;var n=!1;try{var i=[7],u=i[e]();u.next=function(){return{done:n=!0}},i[e]=function(){return u},t(i)}catch(t){}return n}},function(t,r,n){"use strict";var e=n(10),o=n(2),i=n(0),u=n(36),c=n(39);e(e.P+e.R,"Promise",{finally:function(t){var r=u(this,o.Promise||i.Promise),n="function"==typeof t;return this.then(n?function(n){return c(r,t()).then(function(){return n})}:t,n?function(n){return c(r,t()).then(function(){throw n})}:t)}})},function(t,r,n){"use strict";var e=n(10),o=n(24),i=n(38);e(e.S,"Promise",{try:function(t){var r=o.f(this),n=i(t);return(n.e?r.reject:r.resolve)(n.v),r.promise}})},function(t,r,n){var e=function(){return this||"object"==typeof self&&self}()||Function("return this")(),o=e.regeneratorRuntime&&Object.getOwnPropertyNames(e).indexOf("regeneratorRuntime")>=0,i=o&&e.regeneratorRuntime;if(e.regeneratorRuntime=void 0,t.exports=n(76),o)e.regeneratorRuntime=i;else try{delete e.regeneratorRuntime}catch(t){e.regeneratorRuntime=void 0}},function(t,r){!function(r){"use strict";var n,e=Object.prototype,o=e.hasOwnProperty,i="function"==typeof Symbol?Symbol:{},u=i.iterator||"@@iterator",c=i.asyncIterator||"@@asyncIterator",a=i.toStringTag||"@@toStringTag",s="object"==typeof t,f=r.regeneratorRuntime;if(f)s&&(t.exports=f);else{(f=r.regeneratorRuntime=s?t.exports:{}).wrap=w;var l="suspendedStart",p="suspendedYield",h="executing",v="completed",y={},d={};d[u]=function(){return this};var m=Object.getPrototypeOf,g=m&&m(m(M([])));g&&g!==e&&o.call(g,u)&&(d=g);var x=j.prototype=b.prototype=Object.create(d);O.prototype=x.constructor=j,j.constructor=O,j[a]=O.displayName="GeneratorFunction",f.isGeneratorFunction=function(t){var r="function"==typeof t&&t.constructor;return!!r&&(r===O||"GeneratorFunction"===(r.displayName||r.name))},f.mark=function(t){return Object.setPrototypeOf?Object.setPrototypeOf(t,j):(t.__proto__=j,a in t||(t[a]="GeneratorFunction")),t.prototype=Object.create(x),t},f.awrap=function(t){return{__await:t}},S(L.prototype),L.prototype[c]=function(){return this},f.AsyncIterator=L,f.async=function(t,r,n,e){var o=new L(w(t,r,n,e));return f.isGeneratorFunction(r)?o:o.next().then(function(t){return t.done?t.value:o.next()})},S(x),x[a]="Generator",x[u]=function(){return this},x.toString=function(){return"[object Generator]"},f.keys=function(t){var r=[];for(var n in t)r.push(n);return r.reverse(),function n(){for(;r.length;){var e=r.pop();if(e in t)return n.value=e,n.done=!1,n}return n.done=!0,n}},f.values=M,k.prototype={constructor:k,reset:function(t){if(this.prev=0,this.next=0,this.sent=this._sent=n,this.done=!1,this.delegate=null,this.method="next",this.arg=n,this.tryEntries.forEach(T),!t)for(var r in this)"t"===r.charAt(0)&&o.call(this,r)&&!isNaN(+r.slice(1))&&(this[r]=n)},stop:function(){this.done=!0;var t=this.tryEntries[0].completion;if("throw"===t.type)throw t.arg;return this.rval},dispatchException:function(t){if(this.done)throw t;var r=this;function e(e,o){return c.type="throw",c.arg=t,r.next=e,o&&(r.method="next",r.arg=n),!!o}for(var i=this.tryEntries.length-1;i>=0;--i){var u=this.tryEntries[i],c=u.completion;if("root"===u.tryLoc)return e("end");if(u.tryLoc<=this.prev){var a=o.call(u,"catchLoc"),s=o.call(u,"finallyLoc");if(a&&s){if(this.prev<u.catchLoc)return e(u.catchLoc,!0);if(this.prev<u.finallyLoc)return e(u.finallyLoc)}else if(a){if(this.prev<u.catchLoc)return e(u.catchLoc,!0)}else{if(!s)throw new Error("try statement without catch or finally");if(this.prev<u.finallyLoc)return e(u.finallyLoc)}}}},abrupt:function(t,r){for(var n=this.tryEntries.length-1;n>=0;--n){var e=this.tryEntries[n];if(e.tryLoc<=this.prev&&o.call(e,"finallyLoc")&&this.prev<e.finallyLoc){var i=e;break}}i&&("break"===t||"continue"===t)&&i.tryLoc<=r&&r<=i.finallyLoc&&(i=null);var u=i?i.completion:{};return u.type=t,u.arg=r,i?(this.method="next",this.next=i.finallyLoc,y):this.complete(u)},complete:function(t,r){if("throw"===t.type)throw t.arg;return"break"===t.type||"continue"===t.type?this.next=t.arg:"return"===t.type?(this.rval=this.arg=t.arg,this.method="return",this.next="end"):"normal"===t.type&&r&&(this.next=r),y},finish:function(t){for(var r=this.tryEntries.length-1;r>=0;--r){var n=this.tryEntries[r];if(n.finallyLoc===t)return this.complete(n.completion,n.afterLoc),T(n),y}},catch:function(t){for(var r=this.tryEntries.length-1;r>=0;--r){var n=this.tryEntries[r];if(n.tryLoc===t){var e=n.completion;if("throw"===e.type){var o=e.arg;T(n)}return o}}throw new Error("illegal catch attempt")},delegateYield:function(t,r,e){return this.delegate={iterator:M(t),resultName:r,nextLoc:e},"next"===this.method&&(this.arg=n),y}}}function w(t,r,n,e){var o=r&&r.prototype instanceof b?r:b,i=Object.create(o.prototype),u=new k(e||[]);return i._invoke=function(t,r,n){var e=l;return function(o,i){if(e===h)throw new Error("Generator is already running");if(e===v){if("throw"===o)throw i;return N()}for(n.method=o,n.arg=i;;){var u=n.delegate;if(u){var c=P(u,n);if(c){if(c===y)continue;return c}}if("next"===n.method)n.sent=n._sent=n.arg;else if("throw"===n.method){if(e===l)throw e=v,n.arg;n.dispatchException(n.arg)}else"return"===n.method&&n.abrupt("return",n.arg);e=h;var a=_(t,r,n);if("normal"===a.type){if(e=n.done?v:p,a.arg===y)continue;return{value:a.arg,done:n.done}}"throw"===a.type&&(e=v,n.method="throw",n.arg=a.arg)}}}(t,n,u),i}function _(t,r,n){try{return{type:"normal",arg:t.call(r,n)}}catch(t){return{type:"throw",arg:t}}}function b(){}function O(){}function j(){}function S(t){["next","throw","return"].forEach(function(r){t[r]=function(t){return this._invoke(r,t)}})}function L(t){var r;this._invoke=function(n,e){function i(){return new Promise(function(r,i){!function r(n,e,i,u){var c=_(t[n],t,e);if("throw"!==c.type){var a=c.arg,s=a.value;return s&&"object"==typeof s&&o.call(s,"__await")?Promise.resolve(s.__await).then(function(t){r("next",t,i,u)},function(t){r("throw",t,i,u)}):Promise.resolve(s).then(function(t){a.value=t,i(a)},function(t){return r("throw",t,i,u)})}u(c.arg)}(n,e,r,i)})}return r=r?r.then(i,i):i()}}function P(t,r){var e=t.iterator[r.method];if(e===n){if(r.delegate=null,"throw"===r.method){if(t.iterator.return&&(r.method="return",r.arg=n,P(t,r),"throw"===r.method))return y;r.method="throw",r.arg=new TypeError("The iterator does not provide a 'throw' method")}return y}var o=_(e,t.iterator,r.arg);if("throw"===o.type)return r.method="throw",r.arg=o.arg,r.delegate=null,y;var i=o.arg;return i?i.done?(r[t.resultName]=i.value,r.next=t.nextLoc,"return"!==r.method&&(r.method="next",r.arg=n),r.delegate=null,y):i:(r.method="throw",r.arg=new TypeError("iterator result is not an object"),r.delegate=null,y)}function E(t){var r={tryLoc:t[0]};1 in t&&(r.catchLoc=t[1]),2 in t&&(r.finallyLoc=t[2],r.afterLoc=t[3]),this.tryEntries.push(r)}function T(t){var r=t.completion||{};r.type="normal",delete r.arg,t.completion=r}function k(t){this.tryEntries=[{tryLoc:"root"}],t.forEach(E,this),this.reset(!0)}function M(t){if(t){var r=t[u];if(r)return r.call(t);if("function"==typeof t.next)return t;if(!isNaN(t.length)){var e=-1,i=function r(){for(;++e<t.length;)if(o.call(t,e))return r.value=t[e],r.done=!1,r;return r.value=n,r.done=!0,r};return i.next=i}}return{next:N}}function N(){return{value:n,done:!0}}}(function(){return this||"object"==typeof self&&self}()||Function("return this")())},,,,,,,,,,function(t,r,n){"use strict";n.r(r);var e=n(5),o=n.n(e),i=n(6),u=n.n(i),c=n(88),a=n.n(c),s=n(16),f=n(25),l=n(26);function p(t){return h.apply(this,arguments)}function h(){return(h=Object(s.a)(u.a.mark(function t(r){var n,e,i,c,a,s,p,h=arguments;return u.a.wrap(function(t){for(;;)switch(t.prev=t.next){case 0:if(n=h.length>1&&void 0!==h[1]?h[1]:1e5,e=h.length>2&&void 0!==h[2]?h[2]:null,r){t.next=4;break}return t.abrupt("return",o.a.reject("Server error: No string provided to hash"));case 4:if(i=Object(l.promisify)(f.randomBytes),c=Object(l.promisify)(f.pbkdf2),a=64,s="sha512",p="hex",!e||"string"!=typeof e){t.next=13;break}return t.next=12,Buffer.from(e,p);case 12:e=t.sent;case 13:if(e){t.next=17;break}return t.next=16,i(a);case 16:e=t.sent;case 17:return t.abrupt("return",c(r,e,n,a,s).then(function(t){var r=t.toString(p),o=e.toString(p);return"pbkdf2|".concat(n,"|").concat(r,"|").concat(o)}));case 18:case"end":return t.stop()}},t)}))).apply(this,arguments)}exports.handler=function(){var t=Object(s.a)(u.a.mark(function t(r,n){var e,o;return u.a.wrap(function(t){for(;;)switch(t.prev=t.next){case 0:if("POST"===r.httpMethod){t.next=2;break}return t.abrupt("return",{statusCode:405,body:"Method Not Allowed"});case 2:if((e=JSON.parse(r.body)).email&&e.password){t.next=5;break}return t.abrupt("return",{statusCode:400,body:"Must include email and password params"});case 5:return t.t0=e.email,t.next=8,p(e.password,10);case 8:return t.t1=t.sent,o={email:t.t0,password:t.t1},t.abrupt("return",{statusCode:200,body:a()(o)});case 11:case"end":return t.stop()}},t)}));return function(r,n){return t.apply(this,arguments)}}()},,function(t,r,n){t.exports=n(89)},function(t,r,n){var e=n(2),o=e.JSON||(e.JSON={stringify:JSON.stringify});t.exports=function(t){return o.stringify.apply(o,arguments)}}]));
+!function(t,r){for(var n in r)t[n]=r[n]}(exports,function(t){var r={};function n(e){if(r[e])return r[e].exports;var o=r[e]={i:e,l:!1,exports:{}};return t[e].call(o.exports,o,o.exports,n),o.l=!0,o.exports}return n.m=t,n.c=r,n.d=function(t,r,e){n.o(t,r)||Object.defineProperty(t,r,{enumerable:!0,get:e})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,r){if(1&r&&(t=n(t)),8&r)return t;if(4&r&&"object"==typeof t&&t&&t.__esModule)return t;var e=Object.create(null);if(n.r(e),Object.defineProperty(e,"default",{enumerable:!0,value:t}),2&r&&"string"!=typeof t)for(var o in t)n.d(e,o,function(r){return t[r]}.bind(null,o));return e},n.n=function(t){var r=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(r,"a",r),r},n.o=function(t,r){return Object.prototype.hasOwnProperty.call(t,r)},n.p="",n(n.s=88)}([function(t,r){var n=t.exports="undefined"!=typeof window&&window.Math==Math?window:"undefined"!=typeof self&&self.Math==Math?self:Function("return this")();"number"==typeof __g&&(__g=n)},function(t,r,n){var e=n(32)("wks"),o=n(33),i=n(0).Symbol,u="function"==typeof i;(t.exports=function(t){return e[t]||(e[t]=u&&i[t]||(u?i:o)("Symbol."+t))}).store=e},function(t,r){var n=t.exports={version:"2.6.5"};"number"==typeof __e&&(__e=n)},function(t,r,n){var e=n(8);t.exports=function(t){if(!e(t))throw TypeError(t+" is not an object!");return t}},function(t,r,n){t.exports=n(46)},function(t,r,n){t.exports=n(48)},function(t,r,n){var e=n(15),o=n(30);t.exports=n(9)?function(t,r,n){return e.f(t,r,o(1,n))}:function(t,r,n){return t[r]=n,t}},function(t,r,n){var e=n(0),o=n(2),i=n(13),u=n(6),c=n(16),a=function(t,r,n){var s,f,l,p=t&a.F,h=t&a.G,v=t&a.S,y=t&a.P,d=t&a.B,m=t&a.W,g=h?o:o[r]||(o[r]={}),x=g.prototype,w=h?e:v?e[r]:(e[r]||{}).prototype;for(s in h&&(n=r),n)(f=!p&&w&&void 0!==w[s])&&c(g,s)||(l=f?w[s]:n[s],g[s]=h&&"function"!=typeof w[s]?n[s]:d&&f?i(l,e):m&&w[s]==l?function(t){var r=function(r,n,e){if(this instanceof t){switch(arguments.length){case 0:return new t;case 1:return new t(r);case 2:return new t(r,n)}return new t(r,n,e)}return t.apply(this,arguments)};return r.prototype=t.prototype,r}(l):y&&"function"==typeof l?i(Function.call,l):l,y&&((g.virtual||(g.virtual={}))[s]=l,t&a.R&&x&&!x[s]&&u(x,s,l)))};a.F=1,a.G=2,a.S=4,a.P=8,a.B=16,a.W=32,a.U=64,a.R=128,t.exports=a},function(t,r){t.exports=function(t){return"object"==typeof t?null!==t:"function"==typeof t}},function(t,r,n){t.exports=!n(29)(function(){return 7!=Object.defineProperty({},"a",{get:function(){return 7}}).a})},function(t,r){t.exports={}},function(t,r){var n={}.toString;t.exports=function(t){return n.call(t).slice(8,-1)}},function(t,r,n){"use strict";n.d(r,"a",function(){return u});var e=n(5),o=n.n(e);function i(t,r,n,e,i,u,c){try{var a=t[u](c),s=a.value}catch(t){return void n(t)}a.done?r(s):o.a.resolve(s).then(e,i)}function u(t){return function(){var r=this,n=arguments;return new o.a(function(e,o){var u=t.apply(r,n);function c(t){i(u,e,o,c,a,"next",t)}function a(t){i(u,e,o,c,a,"throw",t)}c(void 0)})}}},function(t,r,n){var e=n(14);t.exports=function(t,r,n){if(e(t),void 0===r)return t;switch(n){case 1:return function(n){return t.call(r,n)};case 2:return function(n,e){return t.call(r,n,e)};case 3:return function(n,e,o){return t.call(r,n,e,o)}}return function(){return t.apply(r,arguments)}}},function(t,r){t.exports=function(t){if("function"!=typeof t)throw TypeError(t+" is not a function!");return t}},function(t,r,n){var e=n(3),o=n(51),i=n(52),u=Object.defineProperty;r.f=n(9)?Object.defineProperty:function(t,r,n){if(e(t),r=i(r,!0),e(n),o)try{return u(t,r,n)}catch(t){}if("get"in n||"set"in n)throw TypeError("Accessors not supported!");return"value"in n&&(t[r]=n.value),t}},function(t,r){var n={}.hasOwnProperty;t.exports=function(t,r){return n.call(t,r)}},function(t,r){var n=Math.ceil,e=Math.floor;t.exports=function(t){return isNaN(t=+t)?0:(t>0?e:n)(t)}},function(t,r){t.exports=function(t){if(null==t)throw TypeError("Can't call method on  "+t);return t}},function(t,r){t.exports=!0},function(t,r,n){var e=n(8),o=n(0).document,i=e(o)&&e(o.createElement);t.exports=function(t){return i?o.createElement(t):{}}},function(t,r,n){var e=n(59),o=n(18);t.exports=function(t){return e(o(t))}},function(t,r,n){var e=n(32)("keys"),o=n(33);t.exports=function(t){return e[t]||(e[t]=o(t))}},function(t,r,n){var e=n(15).f,o=n(16),i=n(1)("toStringTag");t.exports=function(t,r,n){t&&!o(t=n?t:t.prototype,i)&&e(t,i,{configurable:!0,value:r})}},function(t,r,n){"use strict";var e=n(14);function o(t){var r,n;this.promise=new t(function(t,e){if(void 0!==r||void 0!==n)throw TypeError("Bad Promise constructor");r=t,n=e}),this.resolve=e(r),this.reject=e(n)}t.exports.f=function(t){return new o(t)}},function(t,r){t.exports=require("crypto")},function(t,r){t.exports=require("util")},function(t,r,n){"use strict";var e=n(50)(!0);n(28)(String,"String",function(t){this._t=String(t),this._i=0},function(){var t,r=this._t,n=this._i;return n>=r.length?{value:void 0,done:!0}:(t=e(r,n),this._i+=t.length,{value:t,done:!1})})},function(t,r,n){"use strict";var e=n(19),o=n(7),i=n(53),u=n(6),c=n(10),a=n(54),s=n(23),f=n(62),l=n(1)("iterator"),p=!([].keys&&"next"in[].keys()),h=function(){return this};t.exports=function(t,r,n,v,y,d,m){a(n,r,v);var g,x,w,_=function(t){if(!p&&t in S)return S[t];switch(t){case"keys":case"values":return function(){return new n(this,t)}}return function(){return new n(this,t)}},b=r+" Iterator",O="values"==y,j=!1,S=t.prototype,L=S[l]||S["@@iterator"]||y&&S[y],P=L||_(y),E=y?O?_("entries"):P:void 0,T="Array"==r&&S.entries||L;if(T&&(w=f(T.call(new t)))!==Object.prototype&&w.next&&(s(w,b,!0),e||"function"==typeof w[l]||u(w,l,h)),O&&L&&"values"!==L.name&&(j=!0,P=function(){return L.call(this)}),e&&!m||!p&&!j&&S[l]||u(S,l,P),c[r]=P,c[b]=h,y)if(g={values:O?P:_("values"),keys:d?P:_("keys"),entries:E},m)for(x in g)x in S||i(S,x,g[x]);else o(o.P+o.F*(p||j),r,g);return g}},function(t,r){t.exports=function(t){try{return!!t()}catch(t){return!0}}},function(t,r){t.exports=function(t,r){return{enumerable:!(1&t),configurable:!(2&t),writable:!(4&t),value:r}}},function(t,r,n){var e=n(17),o=Math.min;t.exports=function(t){return t>0?o(e(t),9007199254740991):0}},function(t,r,n){var e=n(2),o=n(0),i=o["__core-js_shared__"]||(o["__core-js_shared__"]={});(t.exports=function(t,r){return i[t]||(i[t]=void 0!==r?r:{})})("versions",[]).push({version:e.version,mode:n(19)?"pure":"global",copyright:"© 2019 Denis Pushkarev (zloirock.ru)"})},function(t,r){var n=0,e=Math.random();t.exports=function(t){return"Symbol(".concat(void 0===t?"":t,")_",(++n+e).toString(36))}},function(t,r){t.exports="constructor,hasOwnProperty,isPrototypeOf,propertyIsEnumerable,toLocaleString,toString,valueOf".split(",")},function(t,r,n){var e=n(0).document;t.exports=e&&e.documentElement},function(t,r,n){n(64);for(var e=n(0),o=n(6),i=n(10),u=n(1)("toStringTag"),c="CSSRuleList,CSSStyleDeclaration,CSSValueList,ClientRectList,DOMRectList,DOMStringList,DOMTokenList,DataTransferItemList,FileList,HTMLAllCollection,HTMLCollection,HTMLFormElement,HTMLSelectElement,MediaList,MimeTypeArray,NamedNodeMap,NodeList,PaintRequestList,Plugin,PluginArray,SVGLengthList,SVGNumberList,SVGPathSegList,SVGPointList,SVGStringList,SVGTransformList,SourceBufferList,StyleSheetList,TextTrackCueList,TextTrackList,TouchList".split(","),a=0;a<c.length;a++){var s=c[a],f=e[s],l=f&&f.prototype;l&&!l[u]&&o(l,u,s),i[s]=i.Array}},function(t,r,n){var e=n(11),o=n(1)("toStringTag"),i="Arguments"==e(function(){return arguments}());t.exports=function(t){var r,n,u;return void 0===t?"Undefined":null===t?"Null":"string"==typeof(n=function(t,r){try{return t[r]}catch(t){}}(r=Object(t),o))?n:i?e(r):"Object"==(u=e(r))&&"function"==typeof r.callee?"Arguments":u}},function(t,r,n){var e=n(37),o=n(1)("iterator"),i=n(10);t.exports=n(2).getIteratorMethod=function(t){if(null!=t)return t[o]||t["@@iterator"]||i[e(t)]}},function(t,r,n){var e=n(3),o=n(14),i=n(1)("species");t.exports=function(t,r){var n,u=e(t).constructor;return void 0===u||null==(n=e(u)[i])?r:o(n)}},function(t,r,n){var e,o,i,u=n(13),c=n(72),a=n(35),s=n(20),f=n(0),l=f.process,p=f.setImmediate,h=f.clearImmediate,v=f.MessageChannel,y=f.Dispatch,d=0,m={},g=function(){var t=+this;if(m.hasOwnProperty(t)){var r=m[t];delete m[t],r()}},x=function(t){g.call(t.data)};p&&h||(p=function(t){for(var r=[],n=1;arguments.length>n;)r.push(arguments[n++]);return m[++d]=function(){c("function"==typeof t?t:Function(t),r)},e(d),d},h=function(t){delete m[t]},"process"==n(11)(l)?e=function(t){l.nextTick(u(g,t,1))}:y&&y.now?e=function(t){y.now(u(g,t,1))}:v?(i=(o=new v).port2,o.port1.onmessage=x,e=u(i.postMessage,i,1)):f.addEventListener&&"function"==typeof postMessage&&!f.importScripts?(e=function(t){f.postMessage(t+"","*")},f.addEventListener("message",x,!1)):e="onreadystatechange"in s("script")?function(t){a.appendChild(s("script")).onreadystatechange=function(){a.removeChild(this),g.call(t)}}:function(t){setTimeout(u(g,t,1),0)}),t.exports={set:p,clear:h}},function(t,r){t.exports=function(t){try{return{e:!1,v:t()}}catch(t){return{e:!0,v:t}}}},function(t,r,n){var e=n(3),o=n(8),i=n(24);t.exports=function(t,r){if(e(t),o(r)&&r.constructor===t)return r;var n=i.f(t);return(0,n.resolve)(r),n.promise}},function(t,r,n){t.exports=n(80)},function(t,r,n){t.exports=n(83)},function(t,r,n){"use strict";n(43);n(44);var e=n(4),o=n.n(e),i=n(5),u=n.n(i),c=n(12),a=n(25),s=n(26);function f(t){return l.apply(this,arguments)}function l(){return(l=Object(c.a)(o.a.mark(function t(r){var n,e,i,c,f,l,p,h=arguments;return o.a.wrap(function(t){for(;;)switch(t.prev=t.next){case 0:if(n=h.length>1&&void 0!==h[1]?h[1]:1e5,e=h.length>2&&void 0!==h[2]?h[2]:null,r){t.next=4;break}return t.abrupt("return",u.a.reject("Server error: No string provided to hash"));case 4:if(i=Object(s.promisify)(a.randomBytes),c=Object(s.promisify)(a.pbkdf2),f=64,l="sha512",p="hex",!e||"string"!=typeof e){t.next=13;break}return t.next=12,Buffer.from(e,p);case 12:e=t.sent;case 13:if(e){t.next=17;break}return t.next=16,i(f);case 16:e=t.sent;case 17:return t.abrupt("return",c(r,e,n,f,l).then(function(t){var r=t.toString(p),o=e.toString(p);return"pbkdf2|".concat(n,"|").concat(r,"|").concat(o)}));case 18:case"end":return t.stop()}},t)}))).apply(this,arguments)}n.d(r,"a",function(){return f})},function(t,r,n){var e=function(){return this||"object"==typeof self&&self}()||Function("return this")(),o=e.regeneratorRuntime&&Object.getOwnPropertyNames(e).indexOf("regeneratorRuntime")>=0,i=o&&e.regeneratorRuntime;if(e.regeneratorRuntime=void 0,t.exports=n(47),o)e.regeneratorRuntime=i;else try{delete e.regeneratorRuntime}catch(t){e.regeneratorRuntime=void 0}},function(t,r){!function(r){"use strict";var n,e=Object.prototype,o=e.hasOwnProperty,i="function"==typeof Symbol?Symbol:{},u=i.iterator||"@@iterator",c=i.asyncIterator||"@@asyncIterator",a=i.toStringTag||"@@toStringTag",s="object"==typeof t,f=r.regeneratorRuntime;if(f)s&&(t.exports=f);else{(f=r.regeneratorRuntime=s?t.exports:{}).wrap=w;var l="suspendedStart",p="suspendedYield",h="executing",v="completed",y={},d={};d[u]=function(){return this};var m=Object.getPrototypeOf,g=m&&m(m(M([])));g&&g!==e&&o.call(g,u)&&(d=g);var x=j.prototype=b.prototype=Object.create(d);O.prototype=x.constructor=j,j.constructor=O,j[a]=O.displayName="GeneratorFunction",f.isGeneratorFunction=function(t){var r="function"==typeof t&&t.constructor;return!!r&&(r===O||"GeneratorFunction"===(r.displayName||r.name))},f.mark=function(t){return Object.setPrototypeOf?Object.setPrototypeOf(t,j):(t.__proto__=j,a in t||(t[a]="GeneratorFunction")),t.prototype=Object.create(x),t},f.awrap=function(t){return{__await:t}},S(L.prototype),L.prototype[c]=function(){return this},f.AsyncIterator=L,f.async=function(t,r,n,e){var o=new L(w(t,r,n,e));return f.isGeneratorFunction(r)?o:o.next().then(function(t){return t.done?t.value:o.next()})},S(x),x[a]="Generator",x[u]=function(){return this},x.toString=function(){return"[object Generator]"},f.keys=function(t){var r=[];for(var n in t)r.push(n);return r.reverse(),function n(){for(;r.length;){var e=r.pop();if(e in t)return n.value=e,n.done=!1,n}return n.done=!0,n}},f.values=M,k.prototype={constructor:k,reset:function(t){if(this.prev=0,this.next=0,this.sent=this._sent=n,this.done=!1,this.delegate=null,this.method="next",this.arg=n,this.tryEntries.forEach(T),!t)for(var r in this)"t"===r.charAt(0)&&o.call(this,r)&&!isNaN(+r.slice(1))&&(this[r]=n)},stop:function(){this.done=!0;var t=this.tryEntries[0].completion;if("throw"===t.type)throw t.arg;return this.rval},dispatchException:function(t){if(this.done)throw t;var r=this;function e(e,o){return c.type="throw",c.arg=t,r.next=e,o&&(r.method="next",r.arg=n),!!o}for(var i=this.tryEntries.length-1;i>=0;--i){var u=this.tryEntries[i],c=u.completion;if("root"===u.tryLoc)return e("end");if(u.tryLoc<=this.prev){var a=o.call(u,"catchLoc"),s=o.call(u,"finallyLoc");if(a&&s){if(this.prev<u.catchLoc)return e(u.catchLoc,!0);if(this.prev<u.finallyLoc)return e(u.finallyLoc)}else if(a){if(this.prev<u.catchLoc)return e(u.catchLoc,!0)}else{if(!s)throw new Error("try statement without catch or finally");if(this.prev<u.finallyLoc)return e(u.finallyLoc)}}}},abrupt:function(t,r){for(var n=this.tryEntries.length-1;n>=0;--n){var e=this.tryEntries[n];if(e.tryLoc<=this.prev&&o.call(e,"finallyLoc")&&this.prev<e.finallyLoc){var i=e;break}}i&&("break"===t||"continue"===t)&&i.tryLoc<=r&&r<=i.finallyLoc&&(i=null);var u=i?i.completion:{};return u.type=t,u.arg=r,i?(this.method="next",this.next=i.finallyLoc,y):this.complete(u)},complete:function(t,r){if("throw"===t.type)throw t.arg;return"break"===t.type||"continue"===t.type?this.next=t.arg:"return"===t.type?(this.rval=this.arg=t.arg,this.method="return",this.next="end"):"normal"===t.type&&r&&(this.next=r),y},finish:function(t){for(var r=this.tryEntries.length-1;r>=0;--r){var n=this.tryEntries[r];if(n.finallyLoc===t)return this.complete(n.completion,n.afterLoc),T(n),y}},catch:function(t){for(var r=this.tryEntries.length-1;r>=0;--r){var n=this.tryEntries[r];if(n.tryLoc===t){var e=n.completion;if("throw"===e.type){var o=e.arg;T(n)}return o}}throw new Error("illegal catch attempt")},delegateYield:function(t,r,e){return this.delegate={iterator:M(t),resultName:r,nextLoc:e},"next"===this.method&&(this.arg=n),y}}}function w(t,r,n,e){var o=r&&r.prototype instanceof b?r:b,i=Object.create(o.prototype),u=new k(e||[]);return i._invoke=function(t,r,n){var e=l;return function(o,i){if(e===h)throw new Error("Generator is already running");if(e===v){if("throw"===o)throw i;return A()}for(n.method=o,n.arg=i;;){var u=n.delegate;if(u){var c=P(u,n);if(c){if(c===y)continue;return c}}if("next"===n.method)n.sent=n._sent=n.arg;else if("throw"===n.method){if(e===l)throw e=v,n.arg;n.dispatchException(n.arg)}else"return"===n.method&&n.abrupt("return",n.arg);e=h;var a=_(t,r,n);if("normal"===a.type){if(e=n.done?v:p,a.arg===y)continue;return{value:a.arg,done:n.done}}"throw"===a.type&&(e=v,n.method="throw",n.arg=a.arg)}}}(t,n,u),i}function _(t,r,n){try{return{type:"normal",arg:t.call(r,n)}}catch(t){return{type:"throw",arg:t}}}function b(){}function O(){}function j(){}function S(t){["next","throw","return"].forEach(function(r){t[r]=function(t){return this._invoke(r,t)}})}function L(t){var r;this._invoke=function(n,e){function i(){return new Promise(function(r,i){!function r(n,e,i,u){var c=_(t[n],t,e);if("throw"!==c.type){var a=c.arg,s=a.value;return s&&"object"==typeof s&&o.call(s,"__await")?Promise.resolve(s.__await).then(function(t){r("next",t,i,u)},function(t){r("throw",t,i,u)}):Promise.resolve(s).then(function(t){a.value=t,i(a)},function(t){return r("throw",t,i,u)})}u(c.arg)}(n,e,r,i)})}return r=r?r.then(i,i):i()}}function P(t,r){var e=t.iterator[r.method];if(e===n){if(r.delegate=null,"throw"===r.method){if(t.iterator.return&&(r.method="return",r.arg=n,P(t,r),"throw"===r.method))return y;r.method="throw",r.arg=new TypeError("The iterator does not provide a 'throw' method")}return y}var o=_(e,t.iterator,r.arg);if("throw"===o.type)return r.method="throw",r.arg=o.arg,r.delegate=null,y;var i=o.arg;return i?i.done?(r[t.resultName]=i.value,r.next=t.nextLoc,"return"!==r.method&&(r.method="next",r.arg=n),r.delegate=null,y):i:(r.method="throw",r.arg=new TypeError("iterator result is not an object"),r.delegate=null,y)}function E(t){var r={tryLoc:t[0]};1 in t&&(r.catchLoc=t[1]),2 in t&&(r.finallyLoc=t[2],r.afterLoc=t[3]),this.tryEntries.push(r)}function T(t){var r=t.completion||{};r.type="normal",delete r.arg,t.completion=r}function k(t){this.tryEntries=[{tryLoc:"root"}],t.forEach(E,this),this.reset(!0)}function M(t){if(t){var r=t[u];if(r)return r.call(t);if("function"==typeof t.next)return t;if(!isNaN(t.length)){var e=-1,i=function r(){for(;++e<t.length;)if(o.call(t,e))return r.value=t[e],r.done=!1,r;return r.value=n,r.done=!0,r};return i.next=i}}return{next:A}}function A(){return{value:n,done:!0}}}(function(){return this||"object"==typeof self&&self}()||Function("return this")())},function(t,r,n){n(49),n(27),n(36),n(67),n(78),n(79),t.exports=n(2).Promise},function(t,r){},function(t,r,n){var e=n(17),o=n(18);t.exports=function(t){return function(r,n){var i,u,c=String(o(r)),a=e(n),s=c.length;return a<0||a>=s?t?"":void 0:(i=c.charCodeAt(a))<55296||i>56319||a+1===s||(u=c.charCodeAt(a+1))<56320||u>57343?t?c.charAt(a):i:t?c.slice(a,a+2):u-56320+(i-55296<<10)+65536}}},function(t,r,n){t.exports=!n(9)&&!n(29)(function(){return 7!=Object.defineProperty(n(20)("div"),"a",{get:function(){return 7}}).a})},function(t,r,n){var e=n(8);t.exports=function(t,r){if(!e(t))return t;var n,o;if(r&&"function"==typeof(n=t.toString)&&!e(o=n.call(t)))return o;if("function"==typeof(n=t.valueOf)&&!e(o=n.call(t)))return o;if(!r&&"function"==typeof(n=t.toString)&&!e(o=n.call(t)))return o;throw TypeError("Can't convert object to primitive value")}},function(t,r,n){t.exports=n(6)},function(t,r,n){"use strict";var e=n(55),o=n(30),i=n(23),u={};n(6)(u,n(1)("iterator"),function(){return this}),t.exports=function(t,r,n){t.prototype=e(u,{next:o(1,n)}),i(t,r+" Iterator")}},function(t,r,n){var e=n(3),o=n(56),i=n(34),u=n(22)("IE_PROTO"),c=function(){},a=function(){var t,r=n(20)("iframe"),e=i.length;for(r.style.display="none",n(35).appendChild(r),r.src="javascript:",(t=r.contentWindow.document).open(),t.write("<script>document.F=Object<\/script>"),t.close(),a=t.F;e--;)delete a.prototype[i[e]];return a()};t.exports=Object.create||function(t,r){var n;return null!==t?(c.prototype=e(t),n=new c,c.prototype=null,n[u]=t):n=a(),void 0===r?n:o(n,r)}},function(t,r,n){var e=n(15),o=n(3),i=n(57);t.exports=n(9)?Object.defineProperties:function(t,r){o(t);for(var n,u=i(r),c=u.length,a=0;c>a;)e.f(t,n=u[a++],r[n]);return t}},function(t,r,n){var e=n(58),o=n(34);t.exports=Object.keys||function(t){return e(t,o)}},function(t,r,n){var e=n(16),o=n(21),i=n(60)(!1),u=n(22)("IE_PROTO");t.exports=function(t,r){var n,c=o(t),a=0,s=[];for(n in c)n!=u&&e(c,n)&&s.push(n);for(;r.length>a;)e(c,n=r[a++])&&(~i(s,n)||s.push(n));return s}},function(t,r,n){var e=n(11);t.exports=Object("z").propertyIsEnumerable(0)?Object:function(t){return"String"==e(t)?t.split(""):Object(t)}},function(t,r,n){var e=n(21),o=n(31),i=n(61);t.exports=function(t){return function(r,n,u){var c,a=e(r),s=o(a.length),f=i(u,s);if(t&&n!=n){for(;s>f;)if((c=a[f++])!=c)return!0}else for(;s>f;f++)if((t||f in a)&&a[f]===n)return t||f||0;return!t&&-1}}},function(t,r,n){var e=n(17),o=Math.max,i=Math.min;t.exports=function(t,r){return(t=e(t))<0?o(t+r,0):i(t,r)}},function(t,r,n){var e=n(16),o=n(63),i=n(22)("IE_PROTO"),u=Object.prototype;t.exports=Object.getPrototypeOf||function(t){return t=o(t),e(t,i)?t[i]:"function"==typeof t.constructor&&t instanceof t.constructor?t.constructor.prototype:t instanceof Object?u:null}},function(t,r,n){var e=n(18);t.exports=function(t){return Object(e(t))}},function(t,r,n){"use strict";var e=n(65),o=n(66),i=n(10),u=n(21);t.exports=n(28)(Array,"Array",function(t,r){this._t=u(t),this._i=0,this._k=r},function(){var t=this._t,r=this._k,n=this._i++;return!t||n>=t.length?(this._t=void 0,o(1)):o(0,"keys"==r?n:"values"==r?t[n]:[n,t[n]])},"values"),i.Arguments=i.Array,e("keys"),e("values"),e("entries")},function(t,r){t.exports=function(){}},function(t,r){t.exports=function(t,r){return{value:r,done:!!t}}},function(t,r,n){"use strict";var e,o,i,u,c=n(19),a=n(0),s=n(13),f=n(37),l=n(7),p=n(8),h=n(14),v=n(68),y=n(69),d=n(39),m=n(40).set,g=n(73)(),x=n(24),w=n(41),_=n(74),b=n(42),O=a.TypeError,j=a.process,S=j&&j.versions,L=S&&S.v8||"",P=a.Promise,E="process"==f(j),T=function(){},k=o=x.f,M=!!function(){try{var t=P.resolve(1),r=(t.constructor={})[n(1)("species")]=function(t){t(T,T)};return(E||"function"==typeof PromiseRejectionEvent)&&t.then(T)instanceof r&&0!==L.indexOf("6.6")&&-1===_.indexOf("Chrome/66")}catch(t){}}(),A=function(t){var r;return!(!p(t)||"function"!=typeof(r=t.then))&&r},N=function(t,r){if(!t._n){t._n=!0;var n=t._c;g(function(){for(var e=t._v,o=1==t._s,i=0,u=function(r){var n,i,u,c=o?r.ok:r.fail,a=r.resolve,s=r.reject,f=r.domain;try{c?(o||(2==t._h&&C(t),t._h=1),!0===c?n=e:(f&&f.enter(),n=c(e),f&&(f.exit(),u=!0)),n===r.promise?s(O("Promise-chain cycle")):(i=A(n))?i.call(n,a,s):a(n)):s(e)}catch(t){f&&!u&&f.exit(),s(t)}};n.length>i;)u(n[i++]);t._c=[],t._n=!1,r&&!t._h&&R(t)})}},R=function(t){m.call(a,function(){var r,n,e,o=t._v,i=F(t);if(i&&(r=w(function(){E?j.emit("unhandledRejection",o,t):(n=a.onunhandledrejection)?n({promise:t,reason:o}):(e=a.console)&&e.error&&e.error("Unhandled promise rejection",o)}),t._h=E||F(t)?2:1),t._a=void 0,i&&r.e)throw r.v})},F=function(t){return 1!==t._h&&0===(t._a||t._c).length},C=function(t){m.call(a,function(){var r;E?j.emit("rejectionHandled",t):(r=a.onrejectionhandled)&&r({promise:t,reason:t._v})})},G=function(t){var r=this;r._d||(r._d=!0,(r=r._w||r)._v=t,r._s=2,r._a||(r._a=r._c.slice()),N(r,!0))},I=function(t){var r,n=this;if(!n._d){n._d=!0,n=n._w||n;try{if(n===t)throw O("Promise can't be resolved itself");(r=A(t))?g(function(){var e={_w:n,_d:!1};try{r.call(t,s(I,e,1),s(G,e,1))}catch(t){G.call(e,t)}}):(n._v=t,n._s=1,N(n,!1))}catch(t){G.call({_w:n,_d:!1},t)}}};M||(P=function(t){v(this,P,"Promise","_h"),h(t),e.call(this);try{t(s(I,this,1),s(G,this,1))}catch(t){G.call(this,t)}},(e=function(t){this._c=[],this._a=void 0,this._s=0,this._d=!1,this._v=void 0,this._h=0,this._n=!1}).prototype=n(75)(P.prototype,{then:function(t,r){var n=k(d(this,P));return n.ok="function"!=typeof t||t,n.fail="function"==typeof r&&r,n.domain=E?j.domain:void 0,this._c.push(n),this._a&&this._a.push(n),this._s&&N(this,!1),n.promise},catch:function(t){return this.then(void 0,t)}}),i=function(){var t=new e;this.promise=t,this.resolve=s(I,t,1),this.reject=s(G,t,1)},x.f=k=function(t){return t===P||t===u?new i(t):o(t)}),l(l.G+l.W+l.F*!M,{Promise:P}),n(23)(P,"Promise"),n(76)("Promise"),u=n(2).Promise,l(l.S+l.F*!M,"Promise",{reject:function(t){var r=k(this);return(0,r.reject)(t),r.promise}}),l(l.S+l.F*(c||!M),"Promise",{resolve:function(t){return b(c&&this===u?P:this,t)}}),l(l.S+l.F*!(M&&n(77)(function(t){P.all(t).catch(T)})),"Promise",{all:function(t){var r=this,n=k(r),e=n.resolve,o=n.reject,i=w(function(){var n=[],i=0,u=1;y(t,!1,function(t){var c=i++,a=!1;n.push(void 0),u++,r.resolve(t).then(function(t){a||(a=!0,n[c]=t,--u||e(n))},o)}),--u||e(n)});return i.e&&o(i.v),n.promise},race:function(t){var r=this,n=k(r),e=n.reject,o=w(function(){y(t,!1,function(t){r.resolve(t).then(n.resolve,e)})});return o.e&&e(o.v),n.promise}})},function(t,r){t.exports=function(t,r,n,e){if(!(t instanceof r)||void 0!==e&&e in t)throw TypeError(n+": incorrect invocation!");return t}},function(t,r,n){var e=n(13),o=n(70),i=n(71),u=n(3),c=n(31),a=n(38),s={},f={};(r=t.exports=function(t,r,n,l,p){var h,v,y,d,m=p?function(){return t}:a(t),g=e(n,l,r?2:1),x=0;if("function"!=typeof m)throw TypeError(t+" is not iterable!");if(i(m)){for(h=c(t.length);h>x;x++)if((d=r?g(u(v=t[x])[0],v[1]):g(t[x]))===s||d===f)return d}else for(y=m.call(t);!(v=y.next()).done;)if((d=o(y,g,v.value,r))===s||d===f)return d}).BREAK=s,r.RETURN=f},function(t,r,n){var e=n(3);t.exports=function(t,r,n,o){try{return o?r(e(n)[0],n[1]):r(n)}catch(r){var i=t.return;throw void 0!==i&&e(i.call(t)),r}}},function(t,r,n){var e=n(10),o=n(1)("iterator"),i=Array.prototype;t.exports=function(t){return void 0!==t&&(e.Array===t||i[o]===t)}},function(t,r){t.exports=function(t,r,n){var e=void 0===n;switch(r.length){case 0:return e?t():t.call(n);case 1:return e?t(r[0]):t.call(n,r[0]);case 2:return e?t(r[0],r[1]):t.call(n,r[0],r[1]);case 3:return e?t(r[0],r[1],r[2]):t.call(n,r[0],r[1],r[2]);case 4:return e?t(r[0],r[1],r[2],r[3]):t.call(n,r[0],r[1],r[2],r[3])}return t.apply(n,r)}},function(t,r,n){var e=n(0),o=n(40).set,i=e.MutationObserver||e.WebKitMutationObserver,u=e.process,c=e.Promise,a="process"==n(11)(u);t.exports=function(){var t,r,n,s=function(){var e,o;for(a&&(e=u.domain)&&e.exit();t;){o=t.fn,t=t.next;try{o()}catch(e){throw t?n():r=void 0,e}}r=void 0,e&&e.enter()};if(a)n=function(){u.nextTick(s)};else if(!i||e.navigator&&e.navigator.standalone)if(c&&c.resolve){var f=c.resolve(void 0);n=function(){f.then(s)}}else n=function(){o.call(e,s)};else{var l=!0,p=document.createTextNode("");new i(s).observe(p,{characterData:!0}),n=function(){p.data=l=!l}}return function(e){var o={fn:e,next:void 0};r&&(r.next=o),t||(t=o,n()),r=o}}},function(t,r,n){var e=n(0).navigator;t.exports=e&&e.userAgent||""},function(t,r,n){var e=n(6);t.exports=function(t,r,n){for(var o in r)n&&t[o]?t[o]=r[o]:e(t,o,r[o]);return t}},function(t,r,n){"use strict";var e=n(0),o=n(2),i=n(15),u=n(9),c=n(1)("species");t.exports=function(t){var r="function"==typeof o[t]?o[t]:e[t];u&&r&&!r[c]&&i.f(r,c,{configurable:!0,get:function(){return this}})}},function(t,r,n){var e=n(1)("iterator"),o=!1;try{var i=[7][e]();i.return=function(){o=!0},Array.from(i,function(){throw 2})}catch(t){}t.exports=function(t,r){if(!r&&!o)return!1;var n=!1;try{var i=[7],u=i[e]();u.next=function(){return{done:n=!0}},i[e]=function(){return u},t(i)}catch(t){}return n}},function(t,r,n){"use strict";var e=n(7),o=n(2),i=n(0),u=n(39),c=n(42);e(e.P+e.R,"Promise",{finally:function(t){var r=u(this,o.Promise||i.Promise),n="function"==typeof t;return this.then(n?function(n){return c(r,t()).then(function(){return n})}:t,n?function(n){return c(r,t()).then(function(){throw n})}:t)}})},function(t,r,n){"use strict";var e=n(7),o=n(24),i=n(41);e(e.S,"Promise",{try:function(t){var r=o.f(this),n=i(t);return(n.e?r.reject:r.resolve)(n.v),r.promise}})},function(t,r,n){n(81),t.exports=n(2).Array.isArray},function(t,r,n){var e=n(7);e(e.S,"Array",{isArray:n(82)})},function(t,r,n){var e=n(11);t.exports=Array.isArray||function(t){return"Array"==e(t)}},function(t,r,n){n(36),n(27),t.exports=n(84)},function(t,r,n){var e=n(3),o=n(38);t.exports=n(2).getIterator=function(t){var r=o(t);if("function"!=typeof r)throw TypeError(t+" is not iterable!");return e(r.call(t))}},function(t,r,n){t.exports=n(89)},,,function(t,r,n){"use strict";n.r(r);var e=n(4),o=n.n(e),i=n(85),u=n.n(i),c=n(12),a=n(45);exports.handler=function(){var t=Object(c.a)(o.a.mark(function t(r,n){var e,i;return o.a.wrap(function(t){for(;;)switch(t.prev=t.next){case 0:if("POST"===r.httpMethod){t.next=2;break}return t.abrupt("return",{statusCode:405,body:"Method Not Allowed"});case 2:if((e=JSON.parse(r.body)).email&&e.password){t.next=5;break}return t.abrupt("return",{statusCode:400,body:"Must include email and password params"});case 5:return t.t0=e.email,t.next=8,Object(a.a)(e.password,10);case 8:return t.t1=t.sent,i={email:t.t0,password:t.t1},t.abrupt("return",{statusCode:200,body:u()(i)});case 11:case"end":return t.stop()}},t)}));return function(r,n){return t.apply(this,arguments)}}()},function(t,r,n){var e=n(2),o=e.JSON||(e.JSON={stringify:JSON.stringify});t.exports=function(t){return o.stringify.apply(o,arguments)}}]));
