+++ package.json
@@ -16,7 +16,6 @@
     "express": "^4.17.1",
     "netlify-lambda": "^1.4.13",
     "next": "^8.1.0",
-    "querystring": "^0.2.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "styled-components": "^4.3.1"
